rename function name
