+++ superset-frontend/packages/superset-ui-chart-controls/src/utils/index.ts
@@ -26,4 +26,4 @@ export { default as columnChoices } from './columnChoices';
 export * from './defineSavedMetrics';
 export * from './getStandardizedControls';
 export * from './getTemporalColumns';
-export { default as showIfTimeSeries } from './showIfTimeSeries';
+export { default as displayTimeRelatedControls } from './displayTimeRelatedControls';
